Exceptions pt1

+++ src/main/java/com/LSG/dscommerce/services/ProductService.java
@@ -3,14 +3,16 @@ package com.LSG.dscommerce.services;
 import com.LSG.dscommerce.dto.ProductDTO;
 import com.LSG.dscommerce.entities.Product;
 import com.LSG.dscommerce.repositories.ProductRepository;
-import jakarta.persistence.Entity;
+import com.LSG.dscommerce.services.exceptions.DatabaseException;
+import com.LSG.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.Optional;
 
 @Service
 public class ProductService {
@@ -19,10 +21,9 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id){
-        Optional<Product> result = repository.findById(id);
-        Product product = result.get();
-        ProductDTO productDTO = new ProductDTO(product);
-        return productDTO;
+        Product product = repository.findById(id).orElseThrow(
+                ()->new ResourceNotFoundException("Recurso não encontrado"));
+        return new ProductDTO(product);
     }
     @Transactional(readOnly = true)
     public Page<ProductDTO> findAll(Pageable pageable){
@@ -42,20 +43,32 @@ public class ProductService {
     }
     @Transactional
     public ProductDTO update(ProductDTO dto, Long id){
-        Product entity = repository.getReferenceById(id);
+        try{
+            Product entity = repository.getReferenceById(id);
 
-        copyDtoToEntity(dto,entity);
+            copyDtoToEntity(dto,entity);
 
-        entity = repository.save(entity);
+            entity = repository.save(entity);
 
-        return new ProductDTO(entity);
+            return new ProductDTO(entity);
+        }catch (EntityNotFoundException e ){
+            throw new  ResourceNotFoundException("Recurso não encontrado");
+        }
+
     }
 
-    @Transactional()
-    public void delete(Long id){
-        repository.deleteById(id);
+    @Transactional(propagation = Propagation.SUPPORTS)
+    public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
-
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
         entity.setName(dto.getName());
